refactor: Move ExpandedResourceGenerator
